style(controller): separate controller package
