+++ src/main/java/com/e2i/wemeet/web/controller/regist/RegistrationController.java
@@ -1,4 +1,4 @@
-package com.e2i.wemeet.web.controller;
+package com.e2i.wemeet.web.controller.regist;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
